Update some missing fields

+++ src/main/java/discord4j/discordjson/json/gateway/StatusUpdate.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @JsonDeserialize(as = ImmutableStatusUpdate.class)
 public interface StatusUpdate extends PayloadData {
 
-    Optional<Integer> since();
+    Optional<Long> since();
     Optional<ActivityData> game();
     String status();
     boolean afk();
